+++ Assets/Third Party UI Assets/Le Tai's Asset/TranslucentImage/Script/BlurAlgorithm/BlurConfig.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace LeTai.Asset.TranslucentImage {
+    public abstract class BlurConfig : ScriptableObject {
+        public abstract float Strength { get; set; }
+    }
+}