+++ Assets/Third Party UI Assets/MPUIKit/Editor/Scripts/Settings/MPUIKitSettings.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+
+namespace MPUIKIT.Editor {
+    public class MPUIKitSettings : ScriptableObject {
+        [SerializeField] private bool m_SoftMaskSupportInstalled;
+        [SerializeField] private string m_SoftMaskCgincLocation;
+    }
+}